+++ fabric/src/main/java/com/berserker/fabric/FabricMod.java
@@ -10,6 +10,5 @@ public final class FabricMod implements ModInitializer {
         BerserkerClassMod.registerItems();
         BerserkerClassMod.registerSounds();
         BerserkerClassMod.registerEffects();
-        BerserkerClassMod.registerParticles();
     }
 }