+++ neoforge/src/main/java/com/berserker/neoforge/client/NeoForgeClient.java
@@ -6,17 +6,12 @@ import net.neoforged.api.distmarker.Dist;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.fml.event.lifecycle.FMLClientSetupEvent;
-import net.neoforged.neoforge.client.event.RegisterParticleProvidersEvent;
 
 @EventBusSubscriber(modid = BerserkerClassMod.MOD_ID, value = Dist.CLIENT)
 public class NeoForgeClient {
     @SubscribeEvent
     public static void onClientSetup(FMLClientSetupEvent event) {
         BerserkerClient.init();
-    }
-    @SubscribeEvent
-    public static void registerParticleProviders(RegisterParticleProvidersEvent event) {
-        BerserkerClient.registerParticleAppearances();
     }
 }